+++ src/challenges/chess-simp/2022-06.ts
@@ -1,6 +1,9 @@
+import { Board } from '@/board'
 import { Challenge } from '@/challenges/core'
 import { users } from '@/challenges/users'
-import { getCapture } from '@/move'
+import { getCapture, getMoveCoords, getMovePiece, moveIsEqual } from '@/move'
+import { legalMoves_slow } from '@/move/legal'
+import { Color, isBlackPiece, isPawn, isWhitePiece } from '@/piece'
 import { P, match } from 'ts-pattern'
 
 export class Simp_2022_06_03 implements Challenge {
@@ -38,5 +41,41 @@ export class Simp_2022_06_23 implements Challenge {
     // NB: pawns already can only capture if the target is right in front of them, so we don't need a check
     const capture = getCapture(move)
     return !capture || capture.attacker.kingDistance(capture.victim) === 1
+  }
+}
+
+export class Simp_2022_06_21 implements Challenge {
+  meta: Challenge['meta'] = {
+    uuid: '748d2f26-737a-46da-b906-cea510a0f34b',
+    title: 'That Was Frustrating',
+    link: 'https://www.youtube.com/watch?v=wb9JmQMOG7U',
+    challenge:
+      'Chess, but if your piece (not pawn) can get directly behind the enemy piece (not pawn), it has to.',
+    records: new Map([]),
+  }
+
+  private movesBehind(board: Board) {
+    return legalMoves_slow(board).filter((move) => {
+      // Castling never counts so we can just use the primary mover
+      const piece = getMovePiece(board, move)
+      const behind = board.at(getMoveCoords(move).to.s())
+      return isWhitePiece(piece) && !isPawn(piece) && isBlackPiece(behind) && !isPawn(behind)
+    })
+  }
+
+  highlightSquares: Challenge['highlightSquares'] = ({ board }) => {
+    if (board.side === Color.White) {
+      return this.movesBehind(board).map((move) => ({
+        coord: getMoveCoords(move).to,
+        color: 'blue',
+      }))
+    } else {
+      return []
+    }
+  }
+
+  isMoveAllowed: Challenge['isMoveAllowed'] = ({ board, move }) => {
+    const moves = this.movesBehind(board)
+    return moves.length === 0 || moves.some((m) => moveIsEqual(m, move))
   }
 }